server/requirements.txt
asgiref==3.6.0
Brotli==1.0.9
cffi==1.15.1
cryptography==39.0.1
dj-database-url==1.2.0
Django==4.1.7
django-cors-headers==3.14.0
django-environ==0.10.0
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2-binary==2.9.5
pycparser==2.21
pycryptodomex==3.19.0
PyJWT==2.6.0
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
ua-parser==0.10.0
whitenoise==6.3.0
